% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedRefClass.R
\docType{class}
\name{extendedRefClassDefinition-class}
\alias{extendedRefClassDefinition-class}
\title{Extended Reference Definitions}
\description{
The definition for an extended reference class extends the
\code{\link[=ReferenceClasses]{refClassRepresentation}}.  Of note is that the definition holds
the private methods library, the constant methods library, the
static constant environment and the static environment for all
objects of the class.
}
