% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedRefClass.R
\docType{class}
\name{extendedRefGeneratorSlot-class}
\alias{extendedRefGeneratorSlot-class}
\alias{extendedRefGeneratorSlot}
\alias{extendedRefObjectGenerator-class}
\title{Extended Reference Generators}
\description{
A \link[methods:setRefClass]{refObjectGenerator} is turned into a generator
for extended reference classes by replacing the \code{generator} slot
with an \code{extendedRefGeneratorSlot} object and the \code{def} field in the
\code{extendedRefGeneratorSlot} with a \code{extendedRefClassDefinition} object.
}
\details{
The \code{extendedRefGeneratorSlot} adds the \code{static} field holding a
\link{StaticTriad} object to access the static variables and methods without
needing to access them through an instance of the class.
}

