% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsLibrary.R
\docType{class}
\name{MethodsLibrary-class}
\alias{MethodsLibrary-class}
\alias{MethodsLibrary}
\title{Methods Library}
\usage{
MethodsLibrary(methods, method.type, method.parent, parent, .lock)
}
\arguments{
\item{methods}{The list of methods to include in the library.
Methods will be converted to the appropriate type automatically.}

\item{method.type}{The type of method expected, the type should inherit from 'function'.}

\item{method.parent}{The parent environment for methods within the library.
Defaults to the library environment itself.}

\item{parent}{The parent environment for the library.
Defaults to the calling frame.}

\item{.lock}{A logical flag indicating if the library is to be locked and
prevent adding or changing definitions.
Default is to lock if methods are provided.}
}
\description{
A methods library is a container for methods of a specific type.
All objects in the library should be functions of type \code{method.type}.
One exception is made for objects that points back to the
library object, ie. a self referential object.
}
