% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private_methods.R
\docType{class}
\name{PrivateMethod-class}
\alias{PrivateMethod-class}
\alias{privateMethodsLibrary-class}
\alias{privateMethodsLibrary}
\alias{objectPrivateMethods-class}
\alias{private_methods}
\title{Private Methods Library}
\usage{
\S4method{initialize}{privateMethodsLibrary}( methods = list(), parent=NULL
                            , className=NULL, .lock=TRUE)
       \S4method{initialize}{objectPrivateMethods}(thing, library)
}
\arguments{
\item{thing}{an object of class \code{className}.}

\item{library}{a privateMethodsLibrary object.}
}
\description{
Private methods are functions that are intended for internal use.
They do not appear when listing the contents of a reference object.
They are available to the publid methods to call but cannot be accessed
through the \code{object$method()} mechanism that public methods are.

Private methods are stored in a \code{privateMethodsLibrary} object
when stored as part of a class definition.  They are converted to
a \code{objectPrivateMethods} object when attached to a specific instance.
}
