% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedRefClass.R
\name{setExtendedRefClass}
\alias{setExtendedRefClass}
\title{Create an Extended Reference Class.}
\usage{
setExtendedRefClass(Class, contains = character(),
  fields = character(), private = character(), static = character(),
  static.const = list(), methods = list(), private.methods = list(),
  static.methods = list(), where = topenv(parent.frame()), ...)
}
\arguments{
\item{private}{private variables}

\item{static}{static variables}

\item{static.const}{static constants}

\item{private.methods}{Private Methods}

\item{static.methods}{Static Methods}
}
\description{
Extended Reference classes are reference classes that support
static and private methods and variables.
}
