% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefVector.R
\name{setRefVector}
\alias{setRefVector}
\alias{setRefSet}
\title{Reference Vectors and Reference Sets}
\usage{
setRefVector(element, Class = paste0("ReferenceVector<", element, ">"),
  fields = list(), private = character(), static.const = list(),
  methods = list(), ..., where = topenv(parent.frame()))

setRefSet(element, Class = paste0("ReferenceSet<", element, ">"),
  methods = list(), ..., static.const = list(),
  condition.already.contains = c("message", "warning", "error", "none"),
  where = topenv(parent.frame()))
}
\arguments{
\item{element}{The type of element allowed.}

\item{Class}{The Class name for the vector.}

\item{fields}{see \code{\link[methods:setRefClass]{methods::setRefClass()}}}

\item{private}{Character vector indicating the classes of the private variables.}

\item{static.const}{static constants}

\item{methods}{see \code{\link[methods:setRefClass]{methods::setRefClass()}}}

\item{where}{see \code{\link[methods:setRefClass]{methods::setRefClass()}}}
}
\description{
Reference vectors act similar to lists, but have the additional
restriction that all elements must be of the same type.
Reference Sets, Collections of unique objects.
}
\details{
A reference set is defined as a collection of unique objects.
The difference from reference vectors in that they
do not allow duplicate objects.
}
