% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_methods.R
\docType{class}
\name{StaticMethod-class}
\alias{StaticMethod-class}
\alias{StaticMethods-class}
\alias{static_methods}
\title{Static Methods}
\description{
Static methods denote functions that are intended to act on
class static variables only.  They inherit from refMethodDef,
and use the same initializers and methods.
}
