% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TypedEnvironment.R
\docType{class}
\name{TypedEnvironment-class}
\alias{TypedEnvironment-class}
\title{Typed Environments}
\description{
Typed environments are environments that enforce objects to have a specified class.
They essentially work similar to reference classes but without as much overhead.
They provide the foundation for static and private variables of the
extended reference classes.
}
\section{Slots}{

\describe{
\item{\code{classes}}{A named vector indicating the appropriate classes for objects.}

\item{\code{initializer}}{an optional function that can be specified to initialize the values.}
}}

