% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_variables.R
\docType{class}
\name{classStaticEnv-class}
\alias{classStaticEnv-class}
\alias{new_static_env}
\alias{StaticConstEnv-class}
\alias{static_const}
\title{Static and Static Const Variable Environments}
\arguments{
\item{initialized.name}{Name of the variable which indicates if the environment has been initialized or not.}

\item{initialized.state}{The initial state of the initialized variable named in \code{initialized.name}}

\item{className}{is used in assigning the resulting environment name.}

\item{...}{arguments to initialize.}
}
\description{
Static environments contain variables that are common to all
instances of a class.  Static const variables are accessible to
all class objects but cannot be changed.
}
