% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethodsLibrary.R
\docType{class}
\name{MethodsLibrary-class}
\alias{MethodsLibrary-class}
\alias{MethodsLibrary}
\title{Methods Library}
\description{
A methods library is a container for methods of a specific type.
All objects in the library should be functions of type \code{method.type}.
One exception is made for objects that points back to the
library object, ie. a self referential object.
}
\section{Slots}{

\describe{
\item{\code{method.type}}{A string giving the type of method expected,
the type should inherit from 'function', and
preferrably from \link[methods:refMethodDef-class]{refMethodDef}.}
}}

