% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StaticTriad.R
\docType{class}
\name{StaticTriad-class}
\alias{StaticTriad-class}
\alias{StaticTriad}
\alias{$,StaticTriad-method}
\alias{$<-,StaticTriad-method}
\title{The Static Triad}
\usage{
\S4method{$}{StaticTriad}(x, name)

\S4method{$}{StaticTriad}(x, name) <- value
}
\arguments{
\item{x}{A \code{StaticTriad} Object}

\item{name}{The name of the constant, variable, or method}

\item{value}{The replacement value for variables only.}
}
\description{
The static triad consists of a \code{$const} environment for constants,
a \code{$vars} environment for variables, and a \code{$methods} environment for methods.
}

\section{Methods}{

\itemize{
\item \code{$} is defined to make all variables, constants and
methods visible through the single object.
\item \code{$<-} Does the same but protects constants and methods
from being overwritten.
}
}

